refactor: split rolling source responses
