+++ quality/baseline_report.md
@@ -13,11 +13,11 @@ completion claim.
 ## Generation Identity
 
 - Git branch: `feat/enterprise-risk-refactor-continuation`
-- Git commit: `d03743fbbc102616a92febb5cfa937bbdec87063`
+- Git commit: `6ca94d6157a97936b107178e2bda6a5e52d74880`
 
 ## Current Code Size
 
-- Python source files under `src/`: 215
+- Python source files under `src/`: 216
 - Python test files under `tests/`: 103
 - Python packages under `src/`: 9
 - API entry point route/middleware/handler decorators in `src/app/main.py`: 0
@@ -26,7 +26,6 @@ completion claim.
 
 | Path | Lines | Bytes |
 | --- | --- | --- |
-| src/app/services/rolling_stateful_inputs.py | 302 | 10592 |
 | src/app/services/risk/calculation_orchestrator.py | 301 | 9984 |
 | src/app/services/risk/helpers.py | 297 | 10019 |
 | src/app/services/attribution_exposure_history.py | 280 | 9754 |
@@ -46,6 +45,7 @@ completion claim.
 | src/app/services/attribution_engine.py | 244 | 7513 |
 | src/app/services/rolling_risk_free_dependency.py | 242 | 7826 |
 | src/app/services/concentration/parsing.py | 231 | 7793 |
+| src/app/services/concentration/resolvers.py | 231 | 7566 |
 
 ### Largest Functions And Classes
 
@@ -96,12 +96,12 @@ All checks passed!
 - Ruff format check: passed
 
 ```text
-333 files already formatted
+334 files already formatted
 ```
 - Type checking: passed
 
 ```text
-Success: no issues found in 318 source files
+Success: no issues found in 319 source files
 ```
 - Unit coverage snapshot: passed
 
@@ -113,12 +113,12 @@ Success: no issues found in 318 source files
 ........................................................................ [ 65%]
 ........................................................................ [ 78%]
 ...
-src\app\services\rolling_stateful_inputs.py                   80      1     18      1    98%   102
+src\app\services\rolling_stateful_source_responses.py         45      1     16      1    97%   96
 ------------------------------------------------------------------------------------------------------
-TOTAL                                                       6067     91   1028     73    98%
+TOTAL                                                       6074     91   1028     73    98%
 
-172 files skipped due to complete coverage.
-550 passed in 8.60s
+173 files skipped due to complete coverage.
+550 passed in 8.66s
 ```
 
 ## Complexity And Maintainability Snapshot
@@ -148,7 +148,7 @@ src\app\services\risk\__init__.py - A (100.00)
 - Dependency hygiene: passed
 
 ```text
-Scanning 215 files...
+Scanning 216 files...
 
 Success! No dependency issues found.
 ```
